+++ src/main/java/org/bytedeco/javacv/FrameFilter.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2015 Samuel Audet
+ * Copyright (C) 2015-2018 Samuel Audet
  *
  * Licensed either under the Apache License, Version 2.0, or (at your option)
  * under the terms of the GNU General Public License as published by
@@ -43,6 +43,11 @@ public abstract class FrameFilter implements Closeable {
     protected double frameRate;
     protected double aspectRatio;
 
+    protected String afilters;
+    protected int audioChannels;
+    protected int sampleFormat;
+    protected int sampleRate;
+
     public String getFilters() {
         return filters;
     }
@@ -83,6 +88,27 @@ public abstract class FrameFilter implements Closeable {
     }
     public void setAspectRatio(double aspectRatio) {
         this.aspectRatio = aspectRatio;
+    }
+
+    public int getAudioChannels() {
+        return audioChannels;
+    }
+    public void setAudioChannels(int audioChannels) {
+        this.audioChannels = audioChannels;
+    }
+
+    public int getSampleFormat() {
+        return sampleFormat;
+    }
+    public void setSampleFormat(int sampleFormat) {
+        this.sampleFormat = sampleFormat;
+    }
+
+    public int getSampleRate() {
+        return sampleRate;
+    }
+    public void setSampleRate(int sampleRate) {
+        this.sampleRate = sampleRate;
     }
 
     public static class Exception extends IOException {